ows_tests/requirements.txt
gdal==3.6.2
owslib==0.32.1
pytest==8.3.4
pillow==11.1.0
